instantinsight-db/pyproject.toml
[project]
name = "nl2sql"
version = "0.1.0"
description = "A production-ready RAG (Retrieval-Augmented Generation) system for natural language to SQL conversion using PostgreSQL+pgvector, AWS Bedrock Claude, and LangChain"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-aws>=0.1.0",
    "loguru>=0.7.3",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "psycopg2-binary==2.9.9",
    "anthropic>=0.34.0",
    "python-dotenv",
    "boto3>=1.34.131,<1.35.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.1",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "networkx>=3.0",
    "langchain-anthropic>=0.1.0",
    "pyyaml>=6.0.2",
    "pgvector>=0.2.4",
    "psycopg[binary,pool]>=3.1.0",
    "PyAthena[SQLAlchemy,pandas]>=3.0.0",
    "instructor>=1.9.0",
    "tqdm>=4.66.0",
    "alembic>=1.16.4",
    "sqlglot>=27.6.0",
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "redis>=5.0.1",
    "hiredis>=2.3.0",
    "xxhash>=3.4.1",
    "pre-commit>=4.3.0",
    "ibis-framework[athena,postgres]>=10.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
    "pre-commit",
    "isort",
    "pylint>=3.3.7",
    "ipykernel>=6.29.5",
    "vulture>=2.14",
    "beartype>=0.22.3",
    "icontract>=2.7.1",
    "ruff>=0.8.6",
]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "pytest-cov",
    "pytest-watch",
    "hypothesis",
    "requests>=2.32.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "beartype>=0.22.3",
    "icontract>=2.7.1",
    "ruff>=0.8.6",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D"]
ignore = ["D203", "D212"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pylint.messages_control]
disable = [
    "C",  # Convention - disable all convention checks
    "R",  # Refactoring - disable all refactoring suggestions
    "W0718",  # broad-exception-caught
    "W1203",  # logging-fstring-interpolation
    "E0402",  # import-error
    "W0212",  # protected-access
]

[tool.pylint.format]
max-line-length = 88
